requirements.txt
django==2.2
Pillow==6.0.0
psycopg2==2.8.2
djangorestframework==3.9.2
markdown==3.1       # Markdown support for the browsable API.
django-filter==2.1.0  # Filtering support
requests==2.21.0
